fix: ディレクトリ構成を大幅に変更
