+++ pointRanking/src/commands/decorate/decorateCommand.ts
