+++ pointRanking/src/commands/index.ts
@@ -1,0 +1,2 @@
+export { default as decorateCommand } from "./decorate/decorateCommand";
+export { default as rankingCommand } from "./ranking/rankingCommand";
