+++ pointRanking/src/commands/ranking/myPointsCommand.ts
