+++ pointRanking/src/commands/ranking/totalPointsCommand.ts
