+++ pointRanking/src/edgeFunction/callAddFileSharePointsEdgeFunction.ts
