+++ pointRanking/src/edgeFunction/index.ts
@@ -1,0 +1,4 @@
+export { callAddFileSharePointsEdgeFunction } from "./callAddFileSharePointsEdgeFunction";
+export { callAddKeywordPointsEdgeFunction } from "./callAddKeywordPointsEdgeFunction";
+export { callAddMentionPointsEdgeFunction } from "./callAddMentionPointsEdgeFunction";
+export { callAddPointsEdgeFunction } from "./callAddPointsEdgeFunction";
