+++ pointRanking/src/handlers/index.ts
@@ -1,0 +1,3 @@
+export { handleMessageEvent } from "./message/handleMessageEvent";
+export { handleReactionAdded } from "./reaction/handleReactionAdded";
+
