--- pointRanking/src/index.ts
@@ -1,3 +1,0 @@
-import { useSlackbot } from "./hooks/useSlackbot";
-
-console.log("slackbot", useSlackbot());
