+++ pointRanking/src/types/type.ts
